Reorganize Console-related work
Added TicTacToeRuby.Console folder to organize the class files that relate to data reading from and writing to the console.
Renamed a file to match the class name.
