+++ TicTacToeMain.rb
@@ -1,4 +1,4 @@
-require_relative 'IntroductionMessages.rb'
+require_relative 'TicTacToeRuby.Console/Output/game_introduction.rb'
 introduction = GameIntroduction.new()
 message = introduction.welcome_message()
 puts message
